+++ jdaviz/configs/default/plugins/line_lists/line_list_mixin.py
@@ -1,5 +1,3 @@
-import astropy.units as u
-
 from jdaviz.core.events import AddLineListMessage, RedshiftMessage
 
 __all__ = ['LineListMixin']
@@ -20,14 +18,11 @@ class LineListMixin:
         loaded via the notebook.
         """
 
-        # If the helper class has a global redshift (as in Specviz), use it
-        if not isinstance(line_table, str):
-            if hasattr(self, "_redshift") and "redshift" not in line_table.colnames:
-                line_table["redshift"] = u.Quantity(self._redshift)
-
         lt = self.app.get_viewer('spectrum-viewer').load_line_list(line_table,
                                                                    replace=replace,
                                                                    return_table=True)
+
+        # TODO: why is the rest of this logic here and not in viewer.load_line_list?
 
         # Preset lists were returning None table despite loading correctly
         if lt is None: